Rename `surrealdb::param` to `surrealdb::opt`
