+++ lib/src/api/method/create.rs
@@ -1,7 +1,7 @@
 use crate::api::method::Content;
 use crate::api::method::Method;
-use crate::api::param::DbResource;
-use crate::api::param::Param;
+use crate::api::opt::DbResource;
+use crate::api::opt::Param;
 use crate::api::Connection;
 use crate::api::Result;
 use crate::api::Router;
