+++ lib/src/api/method/invalidate.rs
@@ -1,5 +1,5 @@
 use crate::api::method::Method;
-use crate::api::param::Param;
+use crate::api::opt::Param;
 use crate::api::Connection;
 use crate::api::Error;
 use crate::api::ExtraFeatures;
