+++ lib/src/api/method/patch.rs
@@ -1,8 +1,8 @@
 use crate::api::method::Method;
-use crate::api::param::DbResource;
-use crate::api::param::Param;
-use crate::api::param::PatchOp;
-use crate::api::param::Range;
+use crate::api::opt::DbResource;
+use crate::api::opt::Param;
+use crate::api::opt::PatchOp;
+use crate::api::opt::Range;
 use crate::api::Connection;
 use crate::api::Result;
 use crate::api::Router;
