+++ lib/src/api/method/tests/protocol.rs
@@ -1,9 +1,9 @@
 use super::server;
-use crate::api::param::from_value;
-use crate::api::param::DbResponse;
-use crate::api::param::Param;
-use crate::api::param::ServerAddrs;
-use crate::api::param::ToServerAddrs;
+use crate::api::opt::from_value;
+use crate::api::opt::DbResponse;
+use crate::api::opt::Param;
+use crate::api::opt::ServerAddrs;
+use crate::api::opt::ToServerAddrs;
 use crate::api::Connection;
 use crate::api::ExtraFeatures;
 use crate::api::Method;
