+++ lib/src/api/method/tests/server.rs
@@ -1,9 +1,9 @@
 use super::types::Credentials;
 use super::types::User;
 use crate::api::method::Method;
-use crate::api::param::from_json;
-use crate::api::param::from_value;
-use crate::api::param::DbResponse;
+use crate::api::opt::from_json;
+use crate::api::opt::from_value;
+use crate::api::opt::DbResponse;
 use crate::api::Route;
 use crate::sql::Array;
 use crate::sql::Value;
