+++ lib/src/api/method/update.rs
@@ -2,10 +2,10 @@ use crate::api::method::Content;
 use crate::api::method::Merge;
 use crate::api::method::Method;
 use crate::api::method::Patch;
-use crate::api::param::DbResource;
-use crate::api::param::Param;
-use crate::api::param::PatchOp;
-use crate::api::param::Range;
+use crate::api::opt::DbResource;
+use crate::api::opt::Param;
+use crate::api::opt::PatchOp;
+use crate::api::opt::Range;
 use crate::api::Connection;
 use crate::api::Result;
 use crate::api::Router;
