+++ lib/src/api/opt/resource.rs
