+++ lib/src/api/opt/server_addrs/indxdb.rs
@@ -1,8 +1,8 @@
 use crate::api::embedded::Db;
 use crate::api::err::Error;
-use crate::api::param::ServerAddrs;
-use crate::api::param::Strict;
-use crate::api::param::ToServerAddrs;
+use crate::api::opt::ServerAddrs;
+use crate::api::opt::Strict;
+use crate::api::opt::ToServerAddrs;
 use crate::api::storage::IndxDb;
 use crate::api::Result;
 use url::Url;
