+++ lib/src/api/protocol/ws/native.rs
@@ -1,12 +1,12 @@
 use super::LOG;
 use super::PATH;
 use crate::api::err::Error;
-use crate::api::param::from_value;
-use crate::api::param::DbResponse;
-use crate::api::param::Param;
-use crate::api::param::ServerAddrs;
+use crate::api::opt::from_value;
+use crate::api::opt::DbResponse;
+use crate::api::opt::Param;
+use crate::api::opt::ServerAddrs;
 #[cfg(any(feature = "native-tls", feature = "rustls"))]
-use crate::api::param::Tls;
+use crate::api::opt::Tls;
 use crate::api::protocol::ws::Client;
 use crate::api::protocol::ws::Response;
 use crate::api::protocol::ws::PING_INTERVAL;
