+++ lib/src/api/protocol/ws/wasm.rs
@@ -1,10 +1,10 @@
 use super::LOG;
 use super::PATH;
 use crate::api::err::Error;
-use crate::api::param::from_value;
-use crate::api::param::DbResponse;
-use crate::api::param::Param;
-use crate::api::param::ServerAddrs;
+use crate::api::opt::from_value;
+use crate::api::opt::DbResponse;
+use crate::api::opt::Param;
+use crate::api::opt::ServerAddrs;
 use crate::api::protocol::ws::Client;
 use crate::api::protocol::ws::Response;
 use crate::api::protocol::ws::PING_INTERVAL;
