+++ lib/tests/api.rs
@@ -5,12 +5,12 @@ use serde::Serialize;
 use serde_json::json;
 use std::borrow::Cow;
 use std::ops::Bound;
-use surrealdb::param::Database;
-use surrealdb::param::Jwt;
-use surrealdb::param::NameSpace;
-use surrealdb::param::PatchOp;
-use surrealdb::param::Root;
-use surrealdb::param::Scope;
+use surrealdb::opt::Database;
+use surrealdb::opt::Jwt;
+use surrealdb::opt::NameSpace;
+use surrealdb::opt::PatchOp;
+use surrealdb::opt::Root;
+use surrealdb::opt::Scope;
 use surrealdb::sql::statements::BeginStatement;
 use surrealdb::sql::statements::CommitStatement;
 use surrealdb::Surreal;
